Compiled example docs + watch mode
