+++ docs/examples/intro.ts
@@ -1,0 +1,17 @@
+// START
+import { Length, Measure, meters, seconds, Time, Velocity } from "safe-units";
+
+const length: Length = Measure.of(30, meters);
+const time: Time = Measure.of(15, seconds);
+const velocity: Velocity = length.over(time);
+
+console.log(length.toString()); // 30 m
+console.log(time.toString()); // 15 s
+console.log(velocity.toString()); // 2 m * s^-1
+
+// @ts-expect-error ERROR: A measure of m*s isn't assignable to a measure of m/s.
+const error: Velocity = length.times(time);
+// END
+
+// Ensure that error is used
+console.log(error);
