+++ docsgen/index.ts
@@ -1,0 +1,3 @@
+import { emitDocs } from "./emit";
+
+emitDocs();
