+++ package.json
@@ -23,13 +23,17 @@
         "build": "tsc -p src",
         "clean": "rimraf dist docs/build",
         "compile:docs": "tsc -p docsgen",
-        "docs": "npm-run-all -s compile:docs node:docs",
+        "compile:examples": "tsc -p docs/examples",
+        "docs": "npm-run-all -s compile:docs compile:examples node:docs",
+        "docs:watch": "npm-run-all -s compile:docs compile:examples node:docs:watch",
         "lint:docs": "eslint --config eslint.config.mjs docsgen",
+        "lint:examples": "eslint --config eslint.config.mjs docs/examples",
         "lint:src": "eslint --config eslint.config.mjs src",
         "lint:dist": "./scripts/check-typings.sh",
         "lint:test": "eslint --config eslint.config.mjs test",
-        "lint": "npm-run-all -p lint:docs lint:src lint:dist",
+        "lint": "npm-run-all -p lint:docs lint:examples lint:src lint:dist lint:test",
         "node:docs": "node dist/docsgen/index",
+        "node:docs:watch": "node dist/docsgen/watch",
         "test:src": "jest --config jest.config.ts",
         "test:types": "tsc -p test",
         "test": "npm-run-all -p test:src test:types",
